big ol' refactor to (over)use composite builds.
The plan is to migrate the current 'release' kts to Gradle tasks, because the kts is flaky and awkward to restart at a specific point. So if I split it into Gradle tasks then I can restart the release process at any particular point, if necessary. To do this requires composite builds...